Use two suites to test the router client
