+++ tests/integrations/client/client_test.go
@@ -20,10 +20,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"math"
-	"math/rand"
 	"os"
 	"path"
-	"reflect"
 	"sort"
 	"strconv"
 	"strings"
@@ -1110,118 +1108,6 @@ func (suite *clientTestSuite) SetupTest() {
 	suite.grpcSvr.DirectlyGetRaftCluster().ResetRegionCache()
 }
 
-func (suite *clientTestSuite) TestGetRegion() {
-	re := suite.Require()
-	regionID := regionIDAllocator.alloc()
-	region := &metapb.Region{
-		Id: regionID,
-		RegionEpoch: &metapb.RegionEpoch{
-			ConfVer: 1,
-			Version: 1,
-		},
-		Peers: peers,
-	}
-	req := &pdpb.RegionHeartbeatRequest{
-		Header: newHeader(),
-		Region: region,
-		Leader: peers[0],
-	}
-	err := suite.regionHeartbeat.Send(req)
-	re.NoError(err)
-	testutil.Eventually(re, func() bool {
-		r, err := suite.client.GetRegion(context.Background(), []byte("a"))
-		re.NoError(err)
-		if r == nil {
-			return false
-		}
-		return reflect.DeepEqual(region, r.Meta) &&
-			reflect.DeepEqual(peers[0], r.Leader) &&
-			r.Buckets == nil
-	})
-	breq := &pdpb.ReportBucketsRequest{
-		Header: newHeader(),
-		Buckets: &metapb.Buckets{
-			RegionId:   regionID,
-			Version:    1,
-			Keys:       [][]byte{[]byte("a"), []byte("z")},
-			PeriodInMs: 2000,
-			Stats: &metapb.BucketStats{
-				ReadBytes:  []uint64{1},
-				ReadKeys:   []uint64{1},
-				ReadQps:    []uint64{1},
-				WriteBytes: []uint64{1},
-				WriteKeys:  []uint64{1},
-				WriteQps:   []uint64{1},
-			},
-		},
-	}
-	re.NoError(suite.reportBucket.Send(breq))
-	testutil.Eventually(re, func() bool {
-		r, err := suite.client.GetRegion(context.Background(), []byte("a"), opt.WithBuckets())
-		re.NoError(err)
-		if r == nil {
-			return false
-		}
-		return r.Buckets != nil
-	})
-	suite.srv.GetRaftCluster().GetOpts().(*config.PersistOptions).SetRegionBucketEnabled(false)
-
-	testutil.Eventually(re, func() bool {
-		r, err := suite.client.GetRegion(context.Background(), []byte("a"), opt.WithBuckets())
-		re.NoError(err)
-		if r == nil {
-			return false
-		}
-		return r.Buckets == nil
-	})
-	suite.srv.GetRaftCluster().GetOpts().(*config.PersistOptions).SetRegionBucketEnabled(true)
-
-	re.NoError(failpoint.Enable("github.com/tikv/pd/server/grpcClientClosed", `return(true)`))
-	re.NoError(failpoint.Enable("github.com/tikv/pd/server/useForwardRequest", `return(true)`))
-	re.NoError(suite.reportBucket.Send(breq))
-	re.Error(suite.reportBucket.RecvMsg(breq))
-	re.NoError(failpoint.Disable("github.com/tikv/pd/server/grpcClientClosed"))
-	re.NoError(failpoint.Disable("github.com/tikv/pd/server/useForwardRequest"))
-}
-
-func (suite *clientTestSuite) TestGetPrevRegion() {
-	re := suite.Require()
-	regionLen := 10
-	regions := make([]*metapb.Region, 0, regionLen)
-	for i := range regionLen {
-		regionID := regionIDAllocator.alloc()
-		r := &metapb.Region{
-			Id: regionID,
-			RegionEpoch: &metapb.RegionEpoch{
-				ConfVer: 1,
-				Version: 1,
-			},
-			StartKey: []byte{byte(i)},
-			EndKey:   []byte{byte(i + 1)},
-			Peers:    peers,
-		}
-		regions = append(regions, r)
-		req := &pdpb.RegionHeartbeatRequest{
-			Header: newHeader(),
-			Region: r,
-			Leader: peers[0],
-		}
-		err := suite.regionHeartbeat.Send(req)
-		re.NoError(err)
-	}
-	for i := range 20 {
-		testutil.Eventually(re, func() bool {
-			r, err := suite.client.GetPrevRegion(context.Background(), []byte{byte(i)})
-			re.NoError(err)
-			if i > 0 && i < regionLen {
-				return reflect.DeepEqual(peers[0], r.Leader) &&
-					reflect.DeepEqual(regions[i-1], r.Meta)
-			}
-			return r == nil
-		})
-	}
-}
-
 func (suite *clientTestSuite) TestScanRegions() {
 	re := suite.Require()
 	regionLen := 10
@@ -1297,127 +1183,6 @@ func (suite *clientTestSuite) TestScanRegions() {
 	check([]byte{100}, nil, 1, nil)
 	check([]byte{1}, []byte{6}, 0, regions[1:6])
 	check([]byte{1}, []byte{6}, 2, regions[1:3])
-}
-
-func (suite *clientTestSuite) TestGetRegionByID() {
-	re := suite.Require()
-	regionID := regionIDAllocator.alloc()
-	region := &metapb.Region{
-		Id: regionID,
-		RegionEpoch: &metapb.RegionEpoch{
-			ConfVer: 1,
-			Version: 1,
-		},
-		Peers: peers,
-	}
-	req := &pdpb.RegionHeartbeatRequest{
-		Header: newHeader(),
-		Region: region,
-		Leader: peers[0],
-	}
-	err := suite.regionHeartbeat.Send(req)
-	re.NoError(err)
-
-	testutil.Eventually(re, func() bool {
-		r, err := suite.client.GetRegionByID(context.Background(), regionID)
-		re.NoError(err)
-		if r == nil {
-			return false
-		}
-		return reflect.DeepEqual(region, r.Meta) &&
-			reflect.DeepEqual(peers[0], r.Leader)
-	})
-
-	// test WithCallerComponent
-	testutil.Eventually(re, func() bool {
-		r, err := suite.client.
-			WithCallerComponent(caller.GetComponent(0)).
-			GetRegionByID(context.Background(), regionID)
-		re.NoError(err)
-		if r == nil {
-			return false
-		}
-		return reflect.DeepEqual(region, r.Meta) &&
-			reflect.DeepEqual(peers[0], r.Leader)
-	})
-}
-
-func (suite *clientTestSuite) TestGetRegionConcurrently() {
-	suite.client.UpdateOption(opt.EnableRouterClient, true)
-	defer suite.client.UpdateOption(opt.EnableRouterClient, false)
-
-	re := suite.Require()
-	ctx, cancel := context.WithCancel(suite.ctx)
-	defer cancel()
-
-	regions := make([]*metapb.Region, 0, 2)
-	for i := range 2 {
-		regionID := regionIDAllocator.alloc()
-		region := &metapb.Region{
-			Id: regionID,
-			RegionEpoch: &metapb.RegionEpoch{
-				ConfVer: 1,
-				Version: 1,
-			},
-			StartKey: []byte{byte(i)},
-			EndKey:   []byte{byte(i + 1)},
-			Peers:    peers,
-		}
-		re.NoError(suite.regionHeartbeat.Send(&pdpb.RegionHeartbeatRequest{
-			Header: newHeader(),
-			Region: region,
-			Leader: peers[0],
-		}))
-		regions = append(regions, region)
-	}
-
-	const concurrency = 1000
-
-	wg := sync.WaitGroup{}
-	wg.Add(concurrency)
-	for range concurrency {
-		go func() {
-			defer wg.Done()
-			switch rand.Intn(3) {
-			case 0:
-				region := regions[0]
-				testutil.Eventually(re, func() bool {
-					r, err := suite.client.GetRegion(ctx, region.GetStartKey())
-					re.NoError(err)
-					if r == nil {
-						return false
-					}
-					return reflect.DeepEqual(region, r.Meta) &&
-						reflect.DeepEqual(peers[0], r.Leader) &&
-						r.Buckets == nil
-				})
-			case 1:
-				testutil.Eventually(re, func() bool {
-					r, err := suite.client.GetPrevRegion(ctx, regions[1].GetStartKey())
-					re.NoError(err)
-					if r == nil {
-						return false
-					}
-					return reflect.DeepEqual(regions[0], r.Meta) &&
-						reflect.DeepEqual(peers[0], r.Leader) &&
-						r.Buckets == nil
-				})
-			case 2:
-				region := regions[0]
-				testutil.Eventually(re, func() bool {
-					r, err := suite.client.GetRegionByID(ctx, region.GetId())
-					re.NoError(err)
-					if r == nil {
-						return false
-					}
-					return reflect.DeepEqual(region, r.Meta) &&
-						reflect.DeepEqual(peers[0], r.Leader) &&
-						r.Buckets == nil
-				})
-			}
-		}()
-	}
-	wg.Wait()
 }
 
 func (suite *clientTestSuite) TestGetStore() {
